Code highlighting working but looks rough
