+++ src/css/Sorting.css
@@ -39,10 +39,29 @@
 }
 
 .code-content{
-    display: flex;
-    position: fixed;
-    right: 0;
-    top: 60%;
     padding: 20px 20px;
-    padding-right: 50px;
+    padding-left: 50px;
+}
+
+
+/* TODO color scheme needs to change */
+.code-wrapper {
+    border: solid 3px rgb(187, 187, 187);
+    margin: 25px 50px;
+    position: fixed;
+    bottom: 75px;
+    right: 0;
+    border-radius: 5px;
+    background-color: rgba(112, 112, 112, 0.385);
+    box-shadow: rgba(0, 0, 0, 0.25) 0px 14px 28px, rgba(0, 0, 0, 0.22) 0px 10px 10px;
+}
+
+.code-content pre {
+    padding: 0;
+    margin: 0;
+}
+
+.highlight {
+    background-color: rgb(255, 255, 0.75);
+    transition: background-color .8s linear;
 }
